Grok/nightowl1/requirements.txt
rich==13.7.0
psutil==5.9.5
pyyaml
jinja2==3.1.2
flask==2.3.2
openai==0.28.1
networkx==3.1
scikit-learn==1.3.0

